requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pypdf2>=3.0.1
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.0.5
langchain-google-genai>=0.0.5
chromadb>=0.4.18
sentence-transformers>=2.2.2
openai>=1.3.7
anthropic>=0.7.7
google-generativeai>=0.3.2
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.0
tiktoken>=0.5.1
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
